refactor server api in own package
