+++ core/src/main/java/de/qabel/core/index/server/ExternalContactsAccessor.kt
@@ -1,4 +1,6 @@
-package de.qabel.core.index
+package de.qabel.core.index.server
+
+import de.qabel.core.index.RawContact
 
 
 interface ExternalContactsAccessor {
