+++ core/src/main/java/de/qabel/core/index/server/SearchEndpoint.kt
@@ -1,5 +1,7 @@
-package de.qabel.core.index
+package de.qabel.core.index.server
 
+import de.qabel.core.index.FieldType
+import de.qabel.core.index.IndexContact
 import org.apache.http.client.methods.HttpUriRequest
 
 internal interface SearchEndpoint : EndpointBase<List<IndexContact>> {
