+++ core/src/main/java/de/qabel/core/index/server/UpdateEndpoint.kt
@@ -1,6 +1,8 @@
-package de.qabel.core.index
+package de.qabel.core.index.server
 
 import de.qabel.core.crypto.QblECPublicKey
+import de.qabel.core.index.UpdateIdentity
+import de.qabel.core.index.UpdateResult
 import org.apache.http.client.methods.HttpUriRequest
 
 internal interface UpdateEndpoint : EndpointBase<UpdateResult> {
